+++ native/src/tests/example_test.cpp
@@ -1,3 +1,0 @@
-#include <gtest/gtest.h>
-
-TEST(TestSuiteName, TestName) { EXPECT_EQ(1, 1); }