+++ RELEASE_NOTES.md
@@ -1,3 +1,8 @@
+# v0.0.2
+
+* Allow calling loaded script multiple times
+* Prepend root dir to project.properties in the build script
+
 # v0.0.1
 
 Initial